backend/requirements.txt
jinja2<4.0.0,>=3.1.4
alembic<2.0.0,>=1.12.1
httpx<1.0.0,>=0.25.1
psycopg[binary]<4.0.0,>=3.1.13
sqlmodel<1.0.0,>=0.0.21
# Pin bcrypt until passlib supports the latest
bcrypt==4.0.1
pydantic-settings<3.0.0,>=2.2.1
sentry-sdk[fastapi]<2.0.0,>=1.40.6
pyjwt<3.0.0,>=2.8.0
langchain
langchain-openai
langchain-huggingface
langchain-community
langchain-core
fastapi[standard]
